Add ModeStub Class * abstracted out commonly used ModelStub to a new class since it is long and used in various test classes
